% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{abs_label}
\alias{abs_label}
\title{Title Absolute value labels}
\usage{
abs_label(breaks)
}
\arguments{
\item{breaks}{nbreaks}
}
\value{
breaks
}
\description{
Title Absolute value labels
}
\examples{
# scale_x_continuous(labels = abs_label)

}
