% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{coolors_scrape}
\alias{coolors_scrape}
\title{Scrapes values from a coolors.co palette}
\usage{
coolors_scrape(url, cat = FALSE)
}
\arguments{
\item{url}{coolors.co link}

\item{cat}{cat}
}
\value{
vector of colors
}
\description{
Scrapes values from a coolors.co palette
}
\examples{
x <- "https://coolors.co/96adc8-d7ffab-fcff6c-d89d6a-6d454c"
coolors_scrape(x, cat = TRUE)
}
