% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_n_counttocases.R
\name{count_to_cases}
\alias{count_to_cases}
\alias{rep_n}
\title{Rep_n}
\usage{
count_to_cases(tbl, count_col = n)

rep_n(tbl, count_col = n)
}
\arguments{
\item{tbl}{a tbl or contingency table}

\item{count_col}{collum where the frequency is}
}
\value{
a tbl
}
\description{
Rep_n
}
