% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{dates}
\alias{dates}
\title{Dates}
\usage{
dates(
  year = TRUE,
  month = TRUE,
  week = TRUE,
  dow = TRUE,
  doy = TRUE,
  decimal = FALSE,
  quarter = FALSE,
  semester = FALSE
)
}
\arguments{
\item{year}{add year (default)}

\item{month}{add month (default)}

\item{week}{add week (default)}

\item{dow}{add day of the week (default)}

\item{doy}{add day of the year (default)}

\item{decimal}{add decimal}

\item{quarter}{add quarter}

\item{semester}{add semester}
}
\value{

}
\description{
Dates
}
