% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{juice}
\alias{juice}
\title{Juice}
\usage{
juice(prepped_rec, new_data = NULL, ...)
}
\arguments{
\item{prepped_rec}{a prepped recipe}

\item{new_data}{new data ?}

\item{...}{passed on to bake such as composition}
}
\value{

}
\description{
Juice
}
