% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{komma}
\alias{komma}
\title{Title}
\usage{
komma(...)
}
\arguments{
\item{...}{passed on to scales::comma_format}
}
\value{
scale
}
\description{
Title
}
\examples{
library(ggplot2)
data(faithful)
faithful \%>\%
  ggplot(aes(eruptions)) +
  geom_histogram() +
  scale_x_continuous(labels = komma()) +
  theme_center() +
  labs(title = "This is a title", subtitle = "This is a subtitle")
}
