% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mset}
\alias{mset}
\title{Title}
\usage{
mset(mode = "C", ...)
}
\arguments{
\item{mode}{Classification of regression (C or R)}

\item{...}{passed on the yardstick::metric_set()}
}
\value{
metric_set ibject
}
\description{
Title
}
\examples{
mset()
}
