% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{parse_genre_JSON}
\alias{parse_genre_JSON}
\title{Title}
\usage{
parse_genre_JSON(x)
}
\arguments{
\item{x}{character vector of JSON like structuse}
}
\value{
a list of character vectors
}
\description{
Title
}
