% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pederlib-package.R
\docType{package}
\name{pederlib-package}
\alias{pederlib}
\alias{pederlib-package}
\title{pederlib: Personal library}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Peder Sebastian Martinsen \email{peder.s.martinsen@gmail.com}

}
\keyword{internal}
