% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{pr_juice}
\alias{pr_juice}
\title{prep and bake}
\usage{
pr_juice(rec, new_data = NULL, ...)
}
\arguments{
\item{rec}{unprepped_rec}

\item{new_data}{new}

\item{...}{pass}
}
\value{
baked tibble
}
\description{
prep and bake
}
