% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{read_csv_europe}
\alias{read_csv_europe}
\title{Read csv from european encoding}
\usage{
read_csv_europe(file, skip = 0, encoding = "CP1252", ...)
}
\arguments{
\item{file}{Path to file}

\item{skip}{Skipping rows}

\item{encoding}{encoding (CP1252 is defalt)}

\item{...}{passed on to vroom::vroom}
}
\value{

}
\description{
Read csv from european encoding
}
\examples{

}
