% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{read_dir}
\alias{read_dir}
\title{Read multiple files in directory}
\usage{
read_dir(directory, name_col = TRUE, fun = readr::read_csv, ...)
}
\arguments{
\item{directory}{directory}

\item{name_col}{add a column of the name of the file called 'file_name'}

\item{fun}{read function / function to apply to the files in directory}

\item{...}{passed to fun}
}
\value{
a list
}
\description{
Read multiple files in directory
}
