% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rmsle}
\alias{rmsle}
\title{root mean square log error}
\usage{
rmsle(data, ...)
}
\arguments{
\item{data}{data}

\item{...}{needs truth and estimate}
}
\value{

}
\description{
root mean square log error
}
