% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{scale_xy_log10}
\alias{scale_xy_log10}
\title{Both x and y scale on log}
\usage{
scale_xy_log10(...)
}
\arguments{
\item{...}{passed on to scale_x_log10 and scale_y_log10}
}
\value{
scales
}
\description{
Both x and y scale on log
}
