% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{startup}
\alias{startup}
\title{Loading packages}
\usage{
startup(type = 1, paralell = TRUE)
}
\arguments{
\item{type}{type startup}

\item{paralell}{if parralell}
}
\value{

}
\description{
Loading packages
}
\examples{

}
