% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{theme_center}
\alias{theme_center}
\title{Title}
\usage{
theme_center(
  base_size = 11,
  strip_text_size = 12,
  strip_text_margin = 5,
  subtitle_size = 13,
  subtitle_margin = 10,
  plot_title_size = 16,
  plot_title_margin = 10,
  strip_color = "gray90",
  ...
)
}
\arguments{
\item{base_size}{x}

\item{strip_text_size}{x}

\item{strip_text_margin}{x}

\item{subtitle_size}{x}

\item{subtitle_margin}{x}

\item{plot_title_size}{x}

\item{plot_title_margin}{x}

\item{...}{x}
}
\value{

}
\description{
#theme_center\n
Based on theme_minimal\n
title, subtitle are centered \n
}
\examples{

}
