% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{theme_pedr}
\alias{theme_pedr}
\title{Theme Peder}
\usage{
theme_pedr(
  base_size = 11,
  strip_text_size = 12,
  strip_text_margin = 5,
  subtitle_size = 13,
  subtitle_margin = 10,
  plot_title_size = 16,
  plot_title_margin = 10,
  strip_color = "gray90",
  ...
)
}
\arguments{
\item{base_size}{b}

\item{strip_text_size}{b}

\item{strip_text_margin}{b}

\item{subtitle_size}{b}

\item{subtitle_margin}{b}

\item{plot_title_size}{b}

\item{plot_title_margin}{b}

\item{...}{x}
}
\value{

}
\description{
Based on theme_minimal\n
Font is \verb{BentonSans Regular}, and must be installed first\n
title, subtitle are centered
}
