% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{wfs_autoplot}
\alias{wfs_autoplot}
\title{autoplot for workflowsets}
\usage{
wfs_autoplot(
  trained_workflowset,
  metric = "roc_auc",
  n = 1,
  all_metrics = TRUE,
  rank_by_metric = FALSE,
  ...
)
}
\arguments{
\item{trained_workflowset}{xx}

\item{metric}{xx}

\item{n}{xx}

\item{all_metrics}{xx}

\item{rank_by_metric}{xx}

\item{...}{xx}
}
\value{

}
\description{
autoplot for workflowsets
}
\examples{

}
