% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{ctrl}
\alias{ctrl}
\title{control function}
\usage{
ctrl(save_workflow = FALSE, ...)
}
\arguments{
\item{save_workflow}{save workflow}

\item{...}{passed on to control race}
}
\value{

}
\description{
control function
}
