% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{parse_geo}
\alias{parse_geo}
\title{Parsing lat and lng}
\usage{
parse_geo(tbl, geo_variable)
}
\arguments{
\item{tbl}{a tibble}

\item{geo_variable}{variable navn with the lng and lat}
}
\value{

}
\description{
extract latitude and longitude for geo_field like GEO3212(31.3112,-88.22)
!!todo: Velge om det er lat eller long først
}
\examples{

}
