% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitemap_lnglat.R
\name{parse_geo}
\alias{parse_geo}
\title{Parsing lat and lng}
\usage{
parse_geo(tbl, geo_variable, first = "latitude")
}
\arguments{
\item{tbl}{a tibble}

\item{geo_variable}{variable navn with the lng and lat}

\item{first}{the first value is latitude or longitude}
}
\value{

}
\description{
extract latitude and longitude for geo_field like GEO3212(31.3112,-88.22)
}
\examples{

}
