% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pederlib-package.R
\docType{package}
\name{pederlib-package}
\alias{pederlib}
\alias{pederlib-package}
\title{pederlib: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
