% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{startup}
\alias{startup}
\title{Loading packages}
\usage{
startup(type = 1, paralell = FALSE)
}
\arguments{
\item{type}{type startup}

\item{paralell}{if parralell}
}
\value{
nothing
}
\description{
Loading packages
}
\examples{

pederlib::startup(1)
}
