% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{startup}
\alias{startup}
\title{Loading packages}
\usage{
startup(..., parallel = FALSE, stan_op = FALSE, quiet = FALSE)
}
\arguments{
\item{...}{1 Normal, 2 modeling, 3 webscraping, 4 psychometrics, 5 mixedmodels}

\item{parallel}{if parallel}

\item{stan_op}{Include stan}

\item{quiet}{quiet}
}
\value{
nothing
}
\description{
type 1 Normal, 2 modeling, 3 webscraping, 4 psychometrics, 5 mixedmodels
}
\examples{

pederlib::startup(1)
}
