% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{sum_fun}
\alias{sum_fun}
\title{This is a summarize function, and return to into a tibble}
\usage{
sum_fun(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{If NA remove}
}
\value{

}
\description{
\itemize{
\item Mode
\item Median
\item Arithmetic mean
\item Harmonic mean
\item Geometric mean
\item SD
\item SE
\item TODO: Generalized means, Interquartile mean,Pythagorean mean
}
}
\examples{
a <- seq(1, 9, 2)

c <- rep(c(a, NA), 10)
d <- c(1, 2, 3, 3, 20, 31)
e <- c(NA, NA, NA, NA, 3, 3, 2)
f <- c(5, 10, 17, 24, 30)
e <- rnorm(1e4)
sum_fun(a)
sum_fun(c, na.rm = TRUE)
sum_fun(d)
sum_fun(e, na.rm = TRUE)
sum_fun(f)
sum_fun(e)
}
