% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{use_split}
\alias{use_split}
\title{Set of for resampling}
\usage{
use_split(data, strata = NULL, resamples = NULL, number_folds = NULL)
}
\arguments{
\item{data}{Data}

\item{strata}{Strata variable-navn}

\item{resamples}{Type of resampling method}

\item{number_folds}{num of resam}
}
\value{
nothing
}
\description{
Set of for resampling
}
\examples{
use_split(mtcars, mpg)
#Returns:
#library(tidymodels)
#
#set.seed(439)
#mtcars_split <-
# initial_split(mtcars, strata = mpg)

#mtcars_train <-
#   training(mtcars_split)
#mtcars_test <-
#   testing(mtcars_split)
#
#
#set.seed(274)
#mtcars_folds <-
#   bootstraps(mtcars_train, strata = mpg, times = 25)

}
